refactor(desktop): move harness auth hook to hooks/ and split details section
Move use-harness-auth-editor.ts out of components/ (components/** is .tsx-only)
into hooks/agents/workflows/, and split HarnessAuthDetailsSection into
Gateway/ApiKey/Cli sibling files to drop under the frontend size threshold.
No behavior change.
