+++ apps/desktop/src/hooks/agents/workflows/use-harness-auth-editor.ts
